+++ RELEASES.md
@@ -47,6 +47,7 @@ Released on TBD (UTC).
 - Improved live engine message processing to ensure unexpected exceptions result in an immediate hard crash rather than continuing without the queue processing messages
 - Improved handling of negative balances in backtests (#2730), thanks @ms32035
 - Improved robustness of cash and margin account locked balance calculations to avoid negative free balance
+- Improved robustness of fill price parsing for Betfair
 - Improved implementation, validations and testing for Rust instruments (#2723, #2733), thanks @nicolad
 - Improved `Currency` equality to use `strcmp` to avoid C pointer comparison issues with `ustr` string interning
 - Improved unsubscribe cleanup(s) for Bybit adapter